+++ lib/alerts/toast_meta.dart
@@ -12,6 +12,7 @@ class ToastMeta {
   Function? action;
   Function? dismiss;
   Duration duration;
+  Map<String, dynamic>? metaData;
   ToastMeta(
       {this.icon,
       required this.title,
@@ -20,91 +21,102 @@ class ToastMeta {
       this.color,
       this.action,
       this.dismiss,
-      this.duration = const Duration(seconds: 2)});
+      this.duration = const Duration(seconds: 2),
+      this.metaData});
 
-  /// ToastMeta.success() is a pre-defined toast alert with a green background
+  /// Pre-defined style for the [ToastNotificationStyleType.SUCCESS] toast alert.
   ToastMeta.success(
       {Color? backgroundColor,
       String? description,
       String? title,
       Duration? duration,
       Function? action,
+      Map<String, dynamic>? metaData,
       icon = const Icon(Icons.check, color: Colors.green, size: 20)})
       : icon = icon,
         title = title ?? "Success",
         description = description ?? "",
         color = backgroundColor ?? Colors.green.shade50,
         duration = duration ?? const Duration(seconds: 5),
-        action = action ?? null,
+        action = action,
         style = 'success',
+        metaData = metaData,
         super();
 
-  /// ToastMeta.warning() is a pre-defined toast alert with a orange background
+  /// Pre-defined style for the [ToastNotificationStyleType.WARNING] toast alert.
   ToastMeta.warning({
     Color? backgroundColor,
     String? description,
     String? title,
     Duration? duration,
     Function? action,
+    Map<String, dynamic>? metaData,
     icon = const Icon(Icons.error_outline, color: Colors.orange, size: 20),
   })  : icon = icon,
         title = "Oops!",
         description = "",
         color = Colors.orange.shade50,
         duration = const Duration(seconds: 6),
-        action = action ?? null,
+        action = action,
         style = 'warning',
+        metaData = metaData,
         super();
 
-  /// ToastMeta.info() is a pre-defined toast alert with a teal background
+  /// Pre-defined style for the [ToastNotificationStyleType.INFO] toast alert.
   ToastMeta.info({
     Color? backgroundColor,
     String? description,
     String? title,
     Duration? duration,
     Function? action,
+    Map<String, dynamic>? metaData,
     icon = const Icon(Icons.info, color: Colors.teal, size: 20),
   })  : icon = icon,
         title = "Info",
         description = "",
         color = Colors.teal.shade50,
         duration = const Duration(seconds: 5),
-        action = action ?? null,
+        action = action,
         style = 'info',
+        metaData = metaData,
         super();
 
-  /// ToastMeta.danger() is a pre-defined toast alert with a red background
+  /// Pre-defined style for the [ToastNotificationStyleType.DANGER] toast alert.
   ToastMeta.danger({
     Color? backgroundColor,
     String? description,
     String? title,
     Duration? duration,
     Function? action,
+    Map<String, dynamic>? metaData,
     icon = const Icon(Icons.warning, color: Colors.redAccent, size: 20),
   })  : icon = icon,
         title = "Oops!",
         description = "",
         color = Colors.red.shade50,
         duration = const Duration(seconds: 7),
-        action = action ?? null,
+        action = action,
         style = 'danger',
+        metaData = metaData,
         super();
 
-  /// ToastMeta.custom() is a pre-defined toast alert with a red background
+  /// Pre-defined style for the [ToastNotificationStyleType.CUSTOM] toast alert.
   ToastMeta.custom({
     Color? backgroundColor,
     String? description,
     String? title,
     Duration? duration,
     Function? action,
+    Map<String, dynamic>? metaData,
     icon = const Icon(Icons.warning, color: Colors.redAccent, size: 20),
   })  : icon = icon,
         title = "",
         description = "",
         color = Colors.red.shade50,
         duration = const Duration(seconds: 7),
-        action = action ?? null,
+        action = action,
         style = 'custom',
+        metaData = metaData,
         super();
 
   /// ToastMeta.copyWith() is used to copy the current toast alert and